pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "virtual-toaster-source"
version = "0.1.0"
description = "Video source daemon for Video Toaster emulation"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "VTS Contributors"}
]
keywords = ["video", "toaster", "amiga", "emulation", "retro"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
    "Topic :: System :: Emulators",
]

dependencies = [
    "av>=10.0.0",
    "numpy>=1.24.0",
    "Pillow>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.0.270",
    "mypy>=1.0.0",
]
gui = [
    "PyQt6>=6.4.0",
]
capture = [
    # Platform-specific capture dependencies
    # "v4l2py>=0.6.0",  # Linux
]

[project.scripts]
vts-daemon = "vtsource.cli:main_daemon"
vts-test-client = "vtsource.cli:main_test_client"

[project.urls]
Homepage = "https://github.com/peoyli/virtual-toaster-source"
Documentation = "https://github.com/peoyli/virtual-toaster-source/docs"
Repository = "https://github.com/peoyli/virtual-toaster-source"
Issues = "https://github.com/peoyli/virtual-toaster-source/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]  # Line length handled by black

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
